# src/webserver/static/api/swagger.json
{
	"basePath": "/api",
	"consumes": [
		"application/json"
	],
	"definitions": {
		"Customer": {
			"properties": {
				"id": {
					"type": "number",
					"format": "double"
				},
				"status": {
					"type": "string",
					"enum": [
						"Active",
						"Inactive"
					]
				},
				"email": {
					"type": "string"
				},
				"firstName": {
					"type": "string"
				},
				"phoneNumbers": {
					"type": "array",
					"items": {
						"type": "string"
					}
				}
			},
			"required": [
				"id",
				"status",
				"email",
				"firstName",
				"phoneNumbers"
			],
			"type": "object"
		},
		"NewCustomer": {
			"properties": {
				"email": {
					"type": "string"
				},
				"firstName": {
					"type": "string"
				},
				"phoneNumbers": {
					"type": "array",
					"items": {
						"type": "string"
					}
				}
			},
			"required": [
				"email",
				"firstName",
				"phoneNumbers"
			],
			"type": "object"
		}
	},
	"info": {
		"title": "TSOA Express",
		"version": "0.1",
		"description": "Starting point for developing an auto-documenting, API-first application using TypeScript and Node.js."
	},
	"paths": {
		"/customer/{customerId}": {
			"get": {
				"operationId": "GetCustomerEndpoint",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/Customer"
						}
					}
				},
				"description": "Query any customer by customerId",
				"tags": [
					"Customer Information"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "customerId",
						"required": true,
						"format": "double",
						"type": "number"
					}
				]
			},
			"delete": {
				"operationId": "DeleteCustomerEndpoint",
				"produces": [
					"application/json"
				],
				"responses": {
					"202": {
						"description": "Deleted"
					}
				},
				"description": "Delete a customer record",
				"tags": [
					"Customer Information"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "customerId",
						"required": true,
						"format": "double",
						"type": "number"
					},
					{
						"in": "header",
						"name": "Authorization",
						"required": true,
						"type": "string"
					}
				]
			}
		},
		"/customer": {
			"post": {
				"operationId": "CreateCustomerEndpoint",
				"produces": [
					"application/json"
				],
				"responses": {
					"201": {
						"description": "Created",
						"schema": {
							"$ref": "#/definitions/Customer"
						}
					}
				},
				"description": "Create a customer record",
				"tags": [
					"Customer Information"
				],
				"security": [],
				"parameters": [
					{
						"in": "body",
						"name": "newCustomer",
						"required": true,
						"schema": {
							"$ref": "#/definitions/NewCustomer"
						}
					}
				]
			}
		},
		"/math/double/{number}": {
			"get": {
				"operationId": "Double",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"type": "number",
							"format": "double"
						}
					}
				},
				"description": "Double the number provided",
				"tags": [
					"Fun with Math"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "number",
						"required": true,
						"format": "double",
						"type": "number"
					}
				]
			}
		},
		"/math/triple/{number}": {
			"get": {
				"operationId": "Triple",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"type": "number",
							"format": "double"
						}
					}
				},
				"description": "Triple the number provided",
				"tags": [
					"Fun with Math"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "number",
						"required": true,
						"format": "double",
						"type": "number"
					}
				]
			}
		},
		"/math/squareroot/{number}": {
			"get": {
				"operationId": "SquareRoot",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"type": "number",
							"format": "double"
						}
					}
				},
				"description": "Returns the square root of the number provided",
				"tags": [
					"Fun with Math"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "number",
						"required": true,
						"format": "double",
						"type": "number"
					}
				]
			}
		}
	},
	"produces": [
		"application/json"
	],
	"swagger": "2.0",
	"securityDefinitions": {
		"auth": {
			"type": "apiKey",
			"name": "authorization_key",
			"in": "header"
		}
	}
}